reorganize cpp code
