+++ tenseal/cpp/context/sealcontext.cpp
@@ -1,4 +1,4 @@
-#include "tenseal/context/sealcontext.h"
+#include "tenseal/cpp/context/sealcontext.h"
 
 #include <memory>
