+++ tenseal/cpp/context/tensealcontext.h
@@ -2,8 +2,8 @@
 #define TENSEAL_TENSEALCONTEXT_H
 
 #include "seal/seal.h"
-#include "tenseal/context/sealcontext.h"
-#include "tenseal/context/tensealencoder.h"
+#include "tenseal/cpp/context/sealcontext.h"
+#include "tenseal/cpp/context/tensealencoder.h"
 #include "tenseal/proto/tensealcontext.pb.h"
 
 namespace tenseal {
