+++ tenseal/cpp/tensors/bfvvector.cpp
@@ -1,12 +1,12 @@
-#include "tenseal/tensors/bfvvector.h"
+#include "tenseal/cpp/tensors/bfvvector.h"
 
 #include <memory>
 #include <vector>
 
 #include "seal/seal.h"
-#include "tenseal/context/tensealcontext.h"
-#include "tenseal/utils/proto.h"
-#include "tenseal/utils/serialization.h"
+#include "tenseal/cpp/context/tensealcontext.h"
+#include "tenseal/cpp/utils/proto.h"
+#include "tenseal/cpp/utils/serialization.h"
 
 using namespace seal;
 using namespace std;
