+++ tenseal/cpp/tensors/ckksvector.cpp
@@ -1,17 +1,15 @@
-
-#include "tenseal/tensors/ckksvector.h"
-
 #include <cmath>
 #include <memory>
 #include <optional>
 #include <vector>
 
 #include "seal/seal.h"
-#include "tenseal/context/tensealcontext.h"
-#include "tenseal/tensors/utils/matrix_ops.h"
-#include "tenseal/tensors/utils/utils.h"
-#include "tenseal/utils/proto.h"
-#include "tenseal/utils/serialization.h"
+#include "tenseal/cpp/tensors/ckksvector.h"
+#include "tenseal/cpp/context/tensealcontext.h"
+#include "tenseal/cpp/tensors/utils/matrix_ops.h"
+#include "tenseal/cpp/tensors/utils/utils.h"
+#include "tenseal/cpp/utils/proto.h"
+#include "tenseal/cpp/utils/serialization.h"
 
 using namespace seal;
 using namespace std;
