+++ tenseal/cpp/tensors/utils/matrix_ops.h
@@ -4,8 +4,8 @@
 #include <memory>
 
 #include "seal/seal.h"
-#include "tenseal/context/tensealcontext.h"
-#include "tenseal/tensors/utils/utils.h"
+#include "tenseal/cpp/context/tensealcontext.h"
+#include "tenseal/cpp/tensors/utils/utils.h"
 
 namespace tenseal {
