+++ tenseal/cpp/utils/scope.h
