+++ tenseal/tensors/BUILD
@@ -16,34 +16,8 @@
 
 package(default_visibility = ["//visibility:public"])
 
-TENSEAL_DEFAULT_INCLUDES = ["."]
-
-TENSEAL_DEFAULT_COPTS = ["-std=c++17"]
-
-cc_library(
-    name = "tenseal_tensors",
-    srcs = [
-        "bfvvector.cpp",
-        "ckksvector.cpp",
-        "utils/matrix_ops.h",
-        "utils/utils.cpp",
-        "utils/utils.h",
-    ],
-    hdrs = [
-        "bfvvector.h",
-        "ckksvector.h",
-    ],
-    copts = TENSEAL_DEFAULT_COPTS,
-    includes = TENSEAL_DEFAULT_INCLUDES,
-    deps = [
-        "//tenseal/context:tenseal_context",
-        "//tenseal/proto:tenseal_cc_proto",
-        "@com_microsoft_seal//:seal",
-    ],
-)
-
 py_library(
-    name = "tenseal_tensors_py",
+    name = "tenseal_tensors",
     srcs = ["__init__.py"],
     data = ["//tenseal:_tenseal_cpp.so"],
     srcs_version = "PY3",
