--- tenseal/utils/BUILD
@@ -1,36 +1,0 @@
-#
-# Copyright 2020 the authors listed in CONTRIBUTORS.md
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-package(default_visibility = ["//visibility:public"])
-
-TENSEAL_DEFAULT_INCLUDES = ["."]
-
-TENSEAL_DEFAULT_COPTS = ["-std=c++17"]
-
-cc_library(
-    name = "tenseal_utils",
-    hdrs = [
-        "proto.h",
-        "scope.h",
-        "serialization.h",
-    ],
-    copts = TENSEAL_DEFAULT_COPTS,
-    includes = TENSEAL_DEFAULT_INCLUDES,
-    deps = [
-        "//tenseal/proto:tenseal_cc_proto",
-        "@com_microsoft_seal//:seal",
-    ],
-)
