+++ tests/cpp/tensealcontext_test.cpp
@@ -1,4 +1,4 @@
-#include "tenseal/context/tensealcontext.h"
+#include "tenseal/cpp/context/tensealcontext.h"
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
