+++ tests/cpp/tensors/bfvvector_test.cpp
@@ -1,4 +1,4 @@
-#include "tenseal/tensors/bfvvector.h"
+#include "tenseal/cpp/tensors/bfvvector.h"
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
