+++ tests/cpp/tensors/ckksvector_test.cpp
@@ -1,4 +1,4 @@
-#include "tenseal/tensors/ckksvector.h"
+#include "tenseal/cpp/tensors/ckksvector.h"
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
